+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 **Added**
 - compute.InitScript model
 - compute.Cluster model
+- compute.Warehouse model
 - secrets.Secret model
 - secrets.SecretScope model